pipelines/sourcing/pyproject.toml
[tool.poetry]
name = "sourcing"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.26.137"
pandas = "^2.0.1"
python-dotenv = "^1.0.0"
dagit = "^1.3.5"
bios = "^0.1.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
